chore: refactor server into `server` crate
